chore: rename oscd-component to oscd-filtered-list
