+++ index.html
@@ -1,29 +1,29 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en-GB">
-<head>
-  <meta charset="utf-8">
-  <style>
-    body {
-      background: #fafafa;
-    }
-  </style>
-</head>
-<body>
-  <div id="demo"></div>
+  <head>
+    <meta charset="utf-8" />
+    <style>
+      body {
+        background: #fafafa;
+      }
+    </style>
+  </head>
+  <body>
+    <div id="demo"></div>
 
-  <script type="module">
-    import { html, render } from 'lit';
-    import './dist/oscd-component.js';
+    <script type="module">
+      import './dist/oscd-filtered-list.js';
+      import { html, render } from 'lit';
 
-    const title = 'Hello owc World!';
-    render(
-      html`
-        <oscd-component .title=${title}>
-          some light-dom
-        </oscd-component>
-      `,
-      document.querySelector('#demo')
-    );
-  </script>
-</body>
+      const title = 'Hello owc World!';
+      render(
+        html`
+          <oscd-filtered-list .title=${title}>
+            some light-dom
+          </oscd-filtered-list>
+        `,
+        document.querySelector('#demo')
+      );
+    </script>
+  </body>
 </html>
