+++ oscd-filtered-list.spec.ts
@@ -1,13 +1,13 @@
 import { expect, fixture } from '@open-wc/testing';
 import { html } from 'lit';
 
-import './oscd-component.js';
-import type { OscdComponent } from './oscd-component.js';
+import './oscd-filtered-list.js';
+import type { OscdComponent } from './oscd-filtered-list.js';
 
 describe('OscdComponent', () => {
   it('has a default title "Hey there" and counter 5', async () => {
     const el = await fixture<OscdComponent>(
-      html`<oscd-component></oscd-component>`
+      html`<oscd-filtered-list></oscd-filtered-list>`
     );
 
     expect(el.title).to.equal('Hey there');
@@ -16,7 +16,7 @@ describe('OscdComponent', () => {
 
   it('increases the counter on button click', async () => {
     const el = await fixture<OscdComponent>(
-      html`<oscd-component></oscd-component>`
+      html`<oscd-filtered-list></oscd-filtered-list>`
     );
     el.shadowRoot!.querySelector('button')!.click();
 
@@ -25,7 +25,7 @@ describe('OscdComponent', () => {
 
   it('can override the title via attribute', async () => {
     const el = await fixture<OscdComponent>(
-      html`<oscd-component title="attribute title"></oscd-component>`
+      html`<oscd-filtered-list title="attribute title"></oscd-filtered-list>`
     );
 
     expect(el.title).to.equal('attribute title');
@@ -33,7 +33,7 @@ describe('OscdComponent', () => {
 
   it('passes the a11y audit', async () => {
     const el = await fixture<OscdComponent>(
-      html`<oscd-component></oscd-component>`
+      html`<oscd-filtered-list></oscd-filtered-list>`
     );
 
     await expect(el).shadowDom.to.be.accessible();
