+++ oscd-filtered-list.test.ts
@@ -3,8 +3,8 @@ import { html } from 'lit';
 
 import { visualDiff } from '@web/test-runner-visual-regression';
 
-import './oscd-component.js';
-import type { OscdComponent } from './oscd-component.js';
+import './oscd-filtered-list.js';
+import type { OscdComponent } from './oscd-filtered-list.js';
 
 const factor = process.env.CI ? 2 : 1;
 
@@ -16,12 +16,15 @@ function timeout(ms: number) {
 
 mocha.timeout(2000 * factor);
 
-describe('oscd-component', () => {
+describe('oscd-filtered-list', () => {
   let element: OscdComponent;
 
   beforeEach(async () => {
     element = await fixture(
-      html`<oscd-component title="Test Title" counter="41"></oscd-component>`
+      html`<oscd-filtered-list
+        title="Test Title"
+        counter="41"
+      ></oscd-filtered-list>`
     );
     document.body.prepend(element);
   });
@@ -29,12 +32,12 @@ describe('oscd-component', () => {
   afterEach(() => element.remove());
 
   it('displays the title and counter', async () => {
-    await visualDiff(element, 'oscd-component');
+    await visualDiff(element, 'oscd-filtered-list');
   });
 
   it('increments the counter on button click', async () => {
     element.shadowRoot?.querySelector('button')?.click();
     await timeout(100);
-    await visualDiff(element, 'oscd-component-incremented');
+    await visualDiff(element, 'oscd-filtered-list-incremented');
   });
 });
