+++ screenshots/Chromium/baseline/oscd-component-incremented.png
