+++ screenshots/Firefox/baseline/oscd-component-incremented.png
