+++ screenshots/Firefox/baseline/oscd-component.png
