+++ screenshots/Webkit/baseline/oscd-component-incremented.png
